Split symbol step into two components Fix a bug where the connection points set to terminals wouldn't reset
